change commandline ui
FILEs no longer assigned by args but a flag `-f` and it's optional and
only optional in pipe.

TERMs no longer assigned by flag `-t` but using args.

This change make more convenient for daily usage due to TERMs are always
necessary but FILEs is optional when using with pipe.